+++ provider/data_source_on_call_role.go
@@ -81,6 +81,7 @@ func dataSourceOnCallRoleRead(ctx context.Context, d *schema.ResourceData, meta 
 	item, _ := items[0].(*client.OnCallRole)
 
 	d.SetId(item.ID)
-
+	d.Set("slug", item.Slug)
+	d.Set("name", item.Name)
 	return nil
 }